Updated FilesJoiner task

+++ tests/dotnet/Wexflow.Tests/Helper.cs
@@ -77,6 +77,7 @@ namespace Wexflow.Tests
 
         public static void DeleteFilesAndFolders(string folder)
         {
+            if (!Directory.Exists(folder)) return;
             DeleteFiles(folder);
 
             foreach (var dir in Directory.GetDirectories(folder))
@@ -87,6 +88,7 @@ namespace Wexflow.Tests
 
         public static void DeleteFiles(string dir)
         {
+            if (!Directory.Exists(dir)) return;
             foreach (var file in Directory.GetFiles(dir))
             {
                 File.Delete(file);
